+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.2.13 (2023-04-18)
+
+### Changed
+
+- Rework ical schedule export to include final events; also improve changing shifts sync
+
+### Fixed
+
+- Fix issue when creating web overrides for TF schedules using a non-UTC timezone
+
 ## v1.2.12 (2023-04-18)
 
 ### Changed
@@ -33,6 +43,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `alert_group_created_signal` when a new Alert Group is created
 
 ## v1.2.10 (2023-04-13)
+
+### Added
+
+- Added mine filter to schedules listing
 
 ### Fixed